refactor accommodation details - logic in app.js to own view model and fix bug in handlebars view preventing property details rendering
